pyproject.toml
[project]
name = "shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "base58>=2.1.1",
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
]
